4 analysis statistic with NEUTRAL <= 8 [62 rows x 8 columns] CONFIRMATION True long           22 False long miss     19 False 0             12 False short miss     8 True short           1 Name: count, dtype: int64

+++ main.py
@@ -3,20 +3,11 @@ from binance.um_futures import UMFutures
 import time
 import datetime
 import requests
-import cryptocompare
 import csv
-
-# https://pypi.org/project/cryptocompare/
-# print(cryptocompare.get_coin_list(format=True)) # coin_list
-# print(cryptocompare.get_price('BTC', 'USD')) # what is BTC price in USD
-# cryptocompare.get_price(['BTC', 'ETH'], ['EUR', 'GBP'])
-# print(cryptocompare.get_historical_price('BTC', 'EUR', datetime.datetime(2017,6,6))) # history cost
-# print(cryptocompare.get_historical_price_hour('BTC', currency='EUR', limit=1))
-
-# print(cryptocompare.get_price('GFY'))
+import os
 
 
-INTERVAL = Interval.INTERVAL_1_HOUR
+INTERVAL = Interval.INTERVAL_1_MINUTE
 
 TELEGRAM_TOKEN = '6813749013:AAHMwGHqaCpC72ttA-WGdLJyvETXRAYpeb4'
 TELEGRAM_CHANNEL = '@tradingviewname1_bot'
@@ -25,7 +16,7 @@ URL = f'https://api.telegram.org/bot{TELEGRAM_TOKEN}'
 # print(requests.get(URL + '/getUpdates').json()) # выясняем ID чата
 ID = 1191242436
 
-FILE_SAVE_PATH = 'data_3_1h.csv'
+FILE_SAVE_PATH = 'data_4_1m.csv'
 
 def send_message(message):
     URL_MESSAGE = URL + f'/sendMessage?chat_id={ID}&text={message}!'
@@ -83,50 +74,43 @@ def first_data():
             time.sleep(0.01)
         except:
             pass
-            # print('first_data error')
-    # print('longs: ', longs, '\nshorts: ', shorts)
     return longs, shorts
 
 
-def make_time_close_days(time_now_tmp):
-    time_close_tmp = int(time_now_tmp.split(':')[0])
+def make_time_close_days(time_now_tmp) -> str:
+    time_close_tmp = int(str(time_now_tmp).split(':')[1])
     if 0 <= time_close_tmp < 30:
-        if len(str(time_close_tmp + 1)) == 2:
-            time_close_tmp = str(time_close_tmp + 1)
-        else: time_close_tmp = '0' + str(time_close_tmp + 1)
-    else:
-        time_close_tmp = '01'
-    return time_now_tmp.replace(time_now_tmp.split(':')[0], time_close_tmp)
+        time_close_tmp = time_close_tmp + 1
+    elif time_close_tmp == 30:
+        time_close_tmp = 1
+    return time_now_tmp.replace(time_now_tmp.split(':')[1], str(time_close_tmp))
 
 
-def make_time_close_hours(time_now_tmp):
-    time_close_tmp = int(time_now_tmp.split(':')[1])
+def make_time_close_hours(time_now_tmp) -> str:
+    time_close_tmp = int(str(time_now_tmp).split(':')[2])
     if 0 <= time_close_tmp < 23:
-        if len(str(time_close_tmp + 1)) == 2:
-            time_close_tmp = str(time_close_tmp + 1)
-        else: time_close_tmp = '0' + str(time_close_tmp + 1)
-    else:
-        time_close_tmp = '00'
-        time_now_tmp = make_time_close_days(str(time_now_tmp))
-    return time_now_tmp.replace(time_now_tmp.split(':')[1], time_close_tmp)
+        time_close_tmp = time_close_tmp + 1
+    elif time_close_tmp == 23:
+        time_close_tmp = 0
+        time_now_tmp = make_time_close_days(time_now_tmp)
+    return time_now_tmp.replace(time_now_tmp.split(':')[2], str(time_close_tmp))
 
 
-def make_time_close_minutes(time_now_tmp):
-    time_close_tmp = int(time_now_tmp.split(':')[2])
+def make_time_close_minutes(time_now_tmp) -> str:
+    time_close_tmp = int(str(time_now_tmp).split(':')[3])
     if 0 <= time_close_tmp < 59:
-        if len(str(time_close_tmp + 1)) == 2:
-            time_close_tmp = str(time_close_tmp + 1)
-        else: time_close_tmp = '0' + str(time_close_tmp + 1)
-    else:
-        time_close_tmp = '00'
-        time_now_tmp = make_time_close_hours(str(time_now_tmp))
-    return time_now_tmp.replace(time_now_tmp.split(':')[2], time_close_tmp)
+        time_close_tmp = time_close_tmp + 1
+    elif time_close_tmp == 59:
+        time_close_tmp = 0
+        time_now_tmp = make_time_close_hours(time_now_tmp)
+    return time_now_tmp.replace(time_now_tmp.split(':')[3], str(time_close_tmp))
 
-
-with open(FILE_SAVE_PATH, 'w') as csvfile:
-    writer_head = csv.DictWriter(csvfile, fieldnames=['MODE', 'SYMBOL', 'TIME-OPEN', 'PRICE-OPEN', 'TIME-CLOSE',
-                                                      'PRICE-CLOSE', 'RESULT', 'CONFIRMATION'])
-    writer_head.writeheader()
+if not os.path.exists(FILE_SAVE_PATH):
+    with open(FILE_SAVE_PATH, 'w') as csvfile:
+        writer_head = csv.DictWriter(csvfile, fieldnames=['MODE', 'SYMBOL', 'TIME-OPEN', 'PRICE-OPEN', 'TIME-CLOSE',
+                                                          'PRICE-CLOSE', 'RESULT', 'CONFIRMATION'])
+        writer_head.writeheader()
+        print(f'file {FILE_SAVE_PATH} is created.\n')
 
 
 print('START')
@@ -136,47 +120,55 @@ first_data()
 def main():
     main_flag = True
     symbol_times = []
-    count_predicts = 0
+    count_predicts, count_loops = 0, 1
     while main_flag:
         print('====================NEW ROUND===================')
         for i in symbols:
             try:
                 time_now = datetime.datetime.now()
                 data = get_data(i)
-                time_now_tmp = f'{time_now.day}:{time_now.hour}:{time_now.minute}:{time_now.second}'
+                print('==========', i, data, '==========', sep='=====')
+                time_now_tmp = f'{time_now.month}:{time_now.day}:{time_now.hour}:{time_now.minute}:{time_now.second}'
                 symbol = data['SYMBOL']
-                symbol_name, currency = separate_symbol_name(symbol)
-                # print(cryptocompare.get_price(symbol_name, currency))
-                # print(cryptocompare.get_historical_price(symbol, timestamp=time_now))
+                # symbol_name, currency = separate_symbol_name(symbol)
 
-
-                if data['RECOMMENDATION'] == 'STRONG_BUY' and symbol not in longs:
-                    # send_message(symbol + ' BUY')
+                if data['RECOMMENDATION'] == 'STRONG_BUY' and symbol not in longs and int(data['NEUTRAL']) <= 8:
+                    send_message(f'{symbol} + BUY\nCOUNT: {count_predicts}\nNEUTRAL: {data["NEUTRAL"]}\nBUY: '
+                                 f'{data["BUY"]}\nSELL: {data["SELL"]}')
                     with open(FILE_SAVE_PATH, 'a') as csvfile:
                         writer = csv.writer(csvfile)
-                        time_close_tmp = make_time_close_hours(time_now_tmp)
+                        time_close_tmp = make_time_close_minutes(time_now_tmp)
                         writer.writerow(['long', symbol, time_now_tmp, None, time_close_tmp, None, None, None])
                     longs.append(symbol)
                     symbol_times.append((symbol, time_now_tmp))
                     count_predicts += 1
+                    time.sleep(0.1)
 
-                if data['RECOMMENDATION'] == 'STRONG_SELL' and symbol not in shorts:
-                    # send_message(symbol + ' SELL')
+                if data['RECOMMENDATION'] == 'STRONG_SELL' and symbol not in shorts and int(data['NEUTRAL']) <= 8:
+                    send_message(f'{symbol} + SELL\nCOUNT: {count_predicts}\nNEUTRAL: {data["NEUTRAL"]}\nBUY: '
+                                 f'{data["BUY"]}\nSELL: {data["SELL"]}')
                     with open(FILE_SAVE_PATH, 'a') as csvfile:
                         writer = csv.writer(csvfile)
-                        time_close_tmp = make_time_close_hours(time_now_tmp)
+                        time_close_tmp = make_time_close_minutes(time_now_tmp)
                         writer.writerow(['short', symbol, time_now_tmp, None, time_close_tmp, None, None, None])
                     shorts.append(symbol)
                     symbol_times.append((symbol, time_now_tmp))
                     count_predicts += 1
+                    time.sleep(0.1)
 
-                time.sleep(0.1)
                 if not (count_predicts % 100): print(count_predicts)
                 if count_predicts == 10000: main_flag = False
-
             except:
                 pass
-# main()
+
+        if not (count_loops % 3):
+            print(f'{longs},\n{shorts},\nare cleared, count_loops = {count_loops}\n')
+            longs.clear()
+            shorts.clear()
+        count_loops += 1
+
+
+main()
